Add a textfile on first launch
To see the FileBridge folder on an iDevice, a file needs to be
present, otherwise it is hidden to the user.

On the first app launch, create a text document which makes the FileBridge
folder appear. Subsequent app launches will not recreate this text file.

The user can remove the text file once a new file has been placed in there.

Also, update the project information along with splitting each swift file
into separate folders to reduce clutter.

Signed-off-by: kingbri <bdashore3@gmail.com>