Add the Pseudo-fs feature
This feature copies a user's documents to the FileBridge folder
so it's easier to transfer files via iTunes.

It basically moves the document root over to FileBridge so iTunes
can easily interface and the user can transfer files without issues.

This commit adds a bunch of UI changes and tweaks to make the app
look a lot better than previously.

Signed-off-by: kingbri <bdashore3@gmail.com>